rename fake_local_model to static_local_model
