+++ merino/curated_recommendations/ml_backends/static_local_model.py
